Parse methods tweaks
